requirements.txt
#All
numpy
gymnasium
stable-baselines3[extra]
torch
pygame
tqdm
filelock
pyyaml
imageio
Pillow
sb3-contrib
#clash_royal
opencv-python
mss
pyautogui
pure-python-adb
requests
yt-dlp